Fix bug: Word links linking to word page drafts
Old content was recently marked with underscores in front of their
file names, to mark them as drafts to be edited, and to hide them
from menus and linking in the application.

Word page links were generated not from filename, but from the
title set in the page's metadata, and as such, broken links to
word page drafts would appear everywhere.

Fixed this by making the word page links use the page filename
instead.